+++ changelog.md
@@ -4,8 +4,15 @@
 
 ## ✨ What's New:
 
-- Updated panorama.
-- Removed Minecraft and Farming Experience textures from `assets/minecraft/textures/gui/title`.
+- The player will not lose their inventory items on death, instead, they will lose hearts that can be recovered by eating golden apples. Inspired by the [Matcha Flavoured](https://modrinth.com/datapack/matcha-flavoured) datapack by [klei_wright](https://modrinth.com/user/klei_wright).
+- Iron golems and phantoms will not spawn anymore.
+- Zombies are now scared of any light source.
+
+## 🛠️ Minor Changes:
+
+- Changed name from Farming Experience Tweaks to Farming Experience Core.
+- Farming Experience Core now detects correctly if the player is running the modpack from CurseForge or Modrinth.
+- A warning message will now display if the player does not have Farming Experience installed.
 
 ***